VBA Scripts & Screenshots of the data
Hello! I uploaded two VBA scripts for this assignment.  The difference between the files is that 'VBA Challenge.vb' contains the code and 'VBA_Challenge.vb' contains the code with the bonus.  I was able to run both codes on the alphabetical_testing file with no issues, it just took a few minutes for the code with the bonus on it.  When I ran the code with the bonus on the multiple_stock_year_Data file, it crashed excel for me. I used the apostrophe edit to null the bonus code and it worked fine so I'm not sure why it won't go through.